Refactor constants/Laminas app setup separate files.
