+++ module/VuFind/tests/bootstrap.php
@@ -1,38 +1,13 @@
 <?php
 
-// Define path to application directory
-defined('APPLICATION_PATH')
-    || define(
-        'APPLICATION_PATH',
-        (getenv('VUFIND_APPLICATION_PATH') ? getenv('VUFIND_APPLICATION_PATH')
-            : dirname(__DIR__) . '/../..')
-    );
-
-// Define application environment
+// Define application environment (default to testing)
 defined('APPLICATION_ENV')
     || define(
         'APPLICATION_ENV',
         (getenv('VUFIND_ENV') ? getenv('VUFIND_ENV') : 'testing')
     );
 
-// Define default search backend identifier
-defined('DEFAULT_SEARCH_BACKEND') || define('DEFAULT_SEARCH_BACKEND', 'Solr');
-
-// Define path to local override directory
-defined('LOCAL_OVERRIDE_DIR')
-    || define(
-        'LOCAL_OVERRIDE_DIR',
-        (getenv('VUFIND_LOCAL_DIR') ? getenv('VUFIND_LOCAL_DIR') : '')
-    );
-
-// Define path to cache directory
-defined('LOCAL_CACHE_DIR')
-    || define(
-        'LOCAL_CACHE_DIR',
-        (getenv('VUFIND_CACHE_DIR')
-            ? getenv('VUFIND_CACHE_DIR')
-            : (strlen(LOCAL_OVERRIDE_DIR) > 0 ? LOCAL_OVERRIDE_DIR . '/cache' : ''))
-    );
+require __DIR__ . '/../../../config/constants.config.php';
 
 chdir(APPLICATION_PATH);
