+++ public/index.php
@@ -1,61 +1,14 @@
 <?php
-// Define path to application directory
-defined('APPLICATION_PATH')
-    || define(
-        'APPLICATION_PATH',
-        (getenv('VUFIND_APPLICATION_PATH') ? getenv('VUFIND_APPLICATION_PATH')
-            : dirname(__DIR__))
-    );
 
 // If the profiler is enabled, set it up now:
 $vufindProfiler = getenv('VUFIND_PROFILER_XHPROF');
 if (!empty($vufindProfiler)) {
-    include APPLICATION_PATH . '/module/VuFind/functions/profiler.php';
+    include __DIR__ . '/../module/VuFind/functions/profiler.php';
     enableVuFindProfiling($vufindProfiler);
 }
 
-// Define application environment
-defined('APPLICATION_ENV')
-    || define(
-        'APPLICATION_ENV',
-        (getenv('VUFIND_ENV') ? getenv('VUFIND_ENV') : 'production')
-    );
-
-// Define default search backend identifier
-defined('DEFAULT_SEARCH_BACKEND') || define('DEFAULT_SEARCH_BACKEND', 'Solr');
-
-// Define path to local override directory
-defined('LOCAL_OVERRIDE_DIR')
-    || define(
-        'LOCAL_OVERRIDE_DIR',
-        (getenv('VUFIND_LOCAL_DIR') ? getenv('VUFIND_LOCAL_DIR') : '')
-    );
-
-// Define path to cache directory
-defined('LOCAL_CACHE_DIR')
-    || define(
-        'LOCAL_CACHE_DIR',
-        (getenv('VUFIND_CACHE_DIR')
-            ? getenv('VUFIND_CACHE_DIR')
-            : (strlen(LOCAL_OVERRIDE_DIR) > 0 ? LOCAL_OVERRIDE_DIR . '/cache' : ''))
-    );
-
-// Save original working directory in case we need to remember our context, then
-// switch to the application directory for convenience:
-define('ORIGINAL_WORKING_DIRECTORY', getcwd());
-chdir(APPLICATION_PATH);
-
-// Composer autoloading
-if (file_exists('vendor/autoload.php')) {
-    $loader = include 'vendor/autoload.php';
-}
-
-if (!class_exists('Laminas\Loader\AutoloaderFactory')) {
-    throw new RuntimeException('Unable to load Laminas autoloader.');
-}
-
 // Run the application!
-$app = Laminas\Mvc\Application::init(require 'config/application.config.php');
+$app = include __DIR__ . '/../config/application.php';
 if (PHP_SAPI === 'cli') {
     return $app->getServiceManager()
         ->get(\VuFindConsole\ConsoleRunner::class)->run();
